Moved reserve_memory_special for Windows to GTest
